Fix KE testing. Fix OT. Use insecure parameters
Previously, KE testing was based on the == operator of nfl.
However the operator produces a polynomial with 0/1 entries
instead of a boolean, and I was assuming the latter behaviour.
Now a function is used to compare polynomials during testing.

With the correct checks, it was noticed that KE was not working
properly. As a quick fix, the modulus q was enlarged to 30 bits.
It should be noted that nfl supports either moduli of 14 or 30
bits, while the paper suggests a modulus of 15 bits.

Finally, the OT was fixed. The implementation of AES being used
is somewhat faulty, so I pad the inputs to a length that is a
multiple of the block size before encrypting them.